ch6: use i16 for Rect fields

+++ src/game.rs
@@ -43,10 +43,10 @@ impl Walk {
 impl Game for WalkTheDog {
     fn draw(&self, renderer: &Renderer) {
         renderer.clear(&Rect {
-            x: 0.0,
-            y: 0.0,
-            width: 600.0,
-            height: HEIGHT as f32,
+            x: 0,
+            y: 0,
+            width: 600,
+            height: HEIGHT,
         });
 
         if let WalkTheDog::Loaded(walk) = self {
@@ -168,9 +168,9 @@ impl RedHatBoy {
     }
 
     fn bounding_box(&self) -> Rect {
-        const X_OFFSET: f32 = 18.0;
-        const Y_OFFSET: f32 = 14.0;
-        const WIDTH_OFFSET: f32 = 28.0;
+        const X_OFFSET: i16 = 18;
+        const Y_OFFSET: i16 = 14;
+        const WIDTH_OFFSET: i16 = 28;
 
         let mut bounding_box = self.destination_box();
         bounding_box.x += X_OFFSET;
@@ -188,10 +188,8 @@ impl RedHatBoy {
     fn destination_box(&self) -> Rect {
         let sprite = self.current_sprite().expect("Cell not found");
         Rect {
-            x: (self.state_machine.context().position.x + sprite.sprite_source_size.x as i16)
-                .into(),
-            y: (self.state_machine.context().position.y + sprite.sprite_source_size.y as i16)
-                .into(),
+            x: (self.state_machine.context().position.x + sprite.sprite_source_size.x).into(),
+            y: (self.state_machine.context().position.y + sprite.sprite_source_size.y).into(),
             width: sprite.frame.w.into(),
             height: sprite.frame.h.into(),
         }
@@ -228,7 +226,7 @@ impl RedHatBoy {
         self.state_machine = self.state_machine.transition(Event::KnockOut);
     }
 
-    fn land_on(&mut self, position: f32) {
+    fn land_on(&mut self, position: i16) {
         self.state_machine = self.state_machine.transition(Event::Land(position));
     }
 
@@ -272,7 +270,7 @@ enum RedHatBoyStateMachine {
 pub enum Event {
     Jump,
     KnockOut,
-    Land(f32),
+    Land(i16),
     Run,
     Slide,
     Update,
@@ -394,8 +392,8 @@ impl Platform {
     }
 
     fn bounding_boxes(&self) -> Vec<Rect> {
-        const X_OFFSET: f32 = 60.0;
-        const END_HEIGHT: f32 = 54.0;
+        const X_OFFSET: i16 = 60;
+        const END_HEIGHT: i16 = 54;
         let destination_box = self.destination_box();
         let bounding_box_one = Rect {
             x: destination_box.x,
@@ -406,7 +404,7 @@ impl Platform {
         let bounding_box_two = Rect {
             x: destination_box.x + X_OFFSET,
             y: destination_box.y,
-            width: destination_box.width - (X_OFFSET * 2.0),
+            width: destination_box.width - (X_OFFSET * 2),
             height: destination_box.height,
         };
         let bounding_box_three = Rect {
